// generated/src/app/api-models/predi-control.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PrediControl {
  runOptionString: string;
  fileType: string;
  controlFileName: string;
  controlFilePath: string;
  securityCode: string;
  insertUser: string;
  insertDatetime: Date;
  insertProcess: string;
  updateUser: string;
  updateDatetime: Date;
  updateProcess: string;
}